reactivate original checks
